moved files to container, dispatched books to app, implemented combine reducers
