+++ src/components/App.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import BooksList from './bookslist';
-import BooksForm from './booksform';
+import BooksList from '../containers/bookslist';
+import BooksForm from '../containers/booksform';
 import '../App.css';
 
 
 const App = props => {
-  const { books } = props;
+  const { books: { books } } = props;
   return (
     <div>
       <BooksList books={books} />
