+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import BookReducer from './book';
+
+const rootReducer = combineReducers({
+  books: BookReducer,
+});
+
+export default rootReducer;
